add local cluster code for price
