+++ scripts/README.md
@@ -1,0 +1,4 @@
+## Steps to run the scripts
+
+yarn install
+npx hardhat run scripts/cluster/changePrice.ts --network localcluster
