+++ scripts/archive/deploy_redis.ts
@@ -1,7 +1,7 @@
 import 'hardhat-deploy-ethers';
 import '@nomiclabs/hardhat-etherscan';
 import { ethers, network } from 'hardhat';
-import verify from '../utils/verify';
+import verify from '../../utils/verify';
 import hre from 'hardhat';
 import * as fs from 'fs';
 
@@ -78,60 +78,67 @@ const config: ChainConfig = configs[network.name]
     } as ChainConfig);
 
 async function main() {
-  // This is deployer script for emergency deployment of only the postagestamp contract with some quick fixes
+  // This is deployer script for emergency deployment of only the redistribution contract with some quick fixes
   let args: string[] = [];
   let waitTime = 6;
-  let currentRedis = '';
-  let currentOracle = '';
   if (network.name == 'mainnet') {
-    // BZZ Token address, minimumBucketDepth, multisig
-    args = ['0xb1C7F17Ed88189Abf269Bf68A3B2Ed83C5276aAe', '16'];
-    currentRedis = '';
-    currentOracle = '';
+    // Staking, Stamps, Oracle, multisig
+    args = [
+      '0x781c6D1f0eaE6F1Da1F604c6cDCcdB8B76428ba7', // old staking
+      '0x30d155478eF27Ab32A1D578BE7b84BC5988aF381',
+      '0x344A2CC7304B32A87EfDC5407cD4bEC7cf98F035',
+      '0xb1C7F17Ed88189Abf269Bf68A3B2Ed83C5276aAe',
+    ];
   } else if (network.name == 'testnet') {
-    args = ['0x2ac3c1d3e24b45c6c310534bc2dd84b5ed576335', '16'];
-    currentRedis = '0x9e3BDb0c69838CC06D85409d4AD6245e54F70F1d';
-    currentOracle = '0xefC5Ead3188402eCC951DB45827F6e0F99B67a25';
+    args = [
+      '0x484fc3388FD1083fbaF8CE12F282df11F1095Ddf',
+      '0xf86b48B65355D292dDE7da8B4ad1913a72ad45C9',
+      '0x17CFdc0Ac0723ef2c9F39D4BC1FFBeD0405FffeD',
+      '0xb1C7F17Ed88189Abf269Bf68A3B2Ed83C5276aAe',
+    ];
   } else if (network.name == 'localhost') {
-    args = ['0x942C6684eB9874C63d4ed26Ab0623F951D253081', '16'];
-    currentRedis = '0xDF64aed195102E644ad6A0204eD5377589b29618';
-    currentOracle = '0xF52458e65b8e3B69d93DD3803d8ef934c75E0022';
+    args = [
+      '0xFC402f08a61Ebfc4e30e1DB6Cfb38b79578709C1',
+      '0x9A2F29598CB0787Aa806Bbfb65B82A9e558945E7',
+      '0xF52458e65b8e3B69d93DD3803d8ef934c75E0022',
+      '0x62cab2b3b55f341f10348720ca18063cdb779ad5',
+    ];
     waitTime = 1;
   }
 
   // Deploy the contract
-  const stampFactory = await ethers.getContractFactory('PostageStamp');
+  const redisFactory = await ethers.getContractFactory('Redistribution');
   console.log('Deploying contract...');
-  const stamp = await stampFactory.deploy(...args);
-  await stamp.deployed();
-  const deploymentReceipt = await stamp.deployTransaction.wait(waitTime);
+  const redis = await redisFactory.deploy(...args);
+  await redis.deployed();
+  console.log(`Deployed contract to: ${redis.address}`);
+  const deploymentReceipt = await redis.deployTransaction.wait(waitTime);
+
+  // Change roles on current stamps contract
+  const postageStampContract = await ethers.getContractAt('PostageStamp', args[1]);
+  const redistributorRole = ethers.utils.keccak256(ethers.utils.toUtf8Bytes('REDISTRIBUTOR_ROLE'));
+  const tx = await postageStampContract.grantRole(redistributorRole, redis.address);
+  console.log('Changed REDISTRIBUTOR ROLE at : ', tx.hash);
+
+  // Change roles on current staking contract
+  const stakingContract = await ethers.getContractAt('StakeRegistry', args[0]);
+  const tx2 = await stakingContract.grantRole(redistributorRole, redis.address);
+  console.log('Changed REDISTRIBUTOR ROLE at : ', tx2.hash);
 
   // Add metadata for Bee Node
   const deployed = await JSON.parse(JSON.stringify(config.deployedData).toString());
-  const stampABI = await require('../artifacts/src/PostageStamp.sol/PostageStamp.json');
-  deployed['contracts']['postageStamp']['abi'] = stampABI.abi;
-  deployed['contracts']['postageStamp']['bytecode'] = stampABI.bytecode.toString();
-  deployed['contracts']['postageStamp']['address'] = stamp.address;
-  deployed['contracts']['postageStamp']['block'] = deploymentReceipt.blockNumber;
-  deployed['contracts']['postageStamp']['url'] = config.url + stamp.address;
-
-  // We need to first deploy this contract and then use this address and deploy with it redistribution
-  // After that we can add here redis role
-
-  // Change roles on current staking contract
-  const redistributorRole = ethers.utils.keccak256(ethers.utils.toUtf8Bytes('REDISTRIBUTOR_ROLE'));
-  const tx2 = await stamp.grantRole(redistributorRole, currentRedis);
-  console.log('Changed REDISTRIBUTOR ROLE at : ', tx2.hash);
-
-  const oracleRole = ethers.utils.keccak256(ethers.utils.toUtf8Bytes('PRICE_ORACLE_ROLE'));
-  const tx3 = await stamp.grantRole(oracleRole, currentOracle);
-  console.log('Changed ORACLE ROLE at : ', tx3.hash);
+  const redisABI = await require('../artifacts/src/Redistribution.sol/Redistribution.json');
+  deployed['contracts']['redistribution']['abi'] = redisABI.abi;
+  deployed['contracts']['redistribution']['bytecode'] = redisABI.bytecode.toString();
+  deployed['contracts']['redistribution']['address'] = redis.address;
+  deployed['contracts']['redistribution']['block'] = deploymentReceipt.blockNumber;
+  deployed['contracts']['redistribution']['url'] = config.url + redis.address;
 
   fs.writeFileSync(config.networkName + '_deployed.json', JSON.stringify(deployed, null, '\t'));
 
   if ((process.env.MAINNET_ETHERSCAN_KEY || process.env.TESTNET_ETHERSCAN_KEY) && network.name != 'localhost') {
     console.log('Verifying...');
-    await verify(stamp.address, args);
+    await verify(redis.address, args);
   }
 }
