+++ scripts/archive/deploy_stamp.ts
@@ -1,7 +1,7 @@
 import 'hardhat-deploy-ethers';
 import '@nomiclabs/hardhat-etherscan';
 import { ethers, network } from 'hardhat';
-import verify from '../utils/verify';
+import verify from '../../utils/verify';
 import hre from 'hardhat';
 import * as fs from 'fs';
 
@@ -78,56 +78,60 @@ const config: ChainConfig = configs[network.name]
     } as ChainConfig);
 
 async function main() {
-  // This is deployer script for emergency deployment of only the stake contract with some quick fixes
+  // This is deployer script for emergency deployment of only the postagestamp contract with some quick fixes
   let args: string[] = [];
   let waitTime = 6;
   let currentRedis = '';
-  let swarmNetworkID = '1';
+  let currentOracle = '';
   if (network.name == 'mainnet') {
-    swarmNetworkID = '1';
-    // BZZ token, SwarmNetworkId, Multisig
-    args = ['0xb1C7F17Ed88189Abf269Bf68A3B2Ed83C5276aAe', swarmNetworkID, '0xb1C7F17Ed88189Abf269Bf68A3B2Ed83C5276aAe'];
+    // BZZ Token address, minimumBucketDepth, multisig
+    args = ['0xb1C7F17Ed88189Abf269Bf68A3B2Ed83C5276aAe', '16'];
     currentRedis = '';
+    currentOracle = '';
   } else if (network.name == 'testnet') {
-    swarmNetworkID = '333';
-    //args = ['0x0b2bbcbe94d5d4bb782713b137c85d29aa609a13', swarmNetworkID, '0xb1C7F17Ed88189Abf269Bf68A3B2Ed83C5276aAe'];
-    // Old gBZZ
-    args = ['0x2ac3c1d3e24b45c6c310534bc2dd84b5ed576335', swarmNetworkID, '0xb1C7F17Ed88189Abf269Bf68A3B2Ed83C5276aAe'];
-    currentRedis = '0x264079eeF0CE42D790e3FA7DF8D0cfA675ef6504';
+    args = ['0x2ac3c1d3e24b45c6c310534bc2dd84b5ed576335', '16'];
+    currentRedis = '0x9e3BDb0c69838CC06D85409d4AD6245e54F70F1d';
+    currentOracle = '0xefC5Ead3188402eCC951DB45827F6e0F99B67a25';
   } else if (network.name == 'localhost') {
-    swarmNetworkID = '0';
-    args = ['0x942C6684eB9874C63d4ed26Ab0623F951D253081', swarmNetworkID, '0x62cab2b3b55f341f10348720ca18063cdb779ad5'];
-    waitTime = 1;
+    args = ['0x942C6684eB9874C63d4ed26Ab0623F951D253081', '16'];
     currentRedis = '0xDF64aed195102E644ad6A0204eD5377589b29618';
+    currentOracle = '0xF52458e65b8e3B69d93DD3803d8ef934c75E0022';
+    waitTime = 1;
   }
 
   // Deploy the contract
-  const stakeFactory = await ethers.getContractFactory('StakeRegistry');
+  const stampFactory = await ethers.getContractFactory('PostageStamp');
   console.log('Deploying contract...');
-  const stake = await stakeFactory.deploy(...args);
-  await stake.deployed();
-  console.log(`Deployed contract to: ${stake.address}`);
-  const deploymentReceipt = await stake.deployTransaction.wait(waitTime);
+  const stamp = await stampFactory.deploy(...args);
+  await stamp.deployed();
+  const deploymentReceipt = await stamp.deployTransaction.wait(waitTime);
 
   // Add metadata for Bee Node
   const deployed = await JSON.parse(JSON.stringify(config.deployedData).toString());
-  const stakeABI = await require('../artifacts/src/Staking.sol/StakeRegistry.json');
-  deployed['contracts']['staking']['abi'] = stakeABI.abi;
-  deployed['contracts']['staking']['bytecode'] = stakeABI.bytecode.toString();
-  deployed['contracts']['staking']['address'] = stake.address;
-  deployed['contracts']['staking']['block'] = deploymentReceipt.blockNumber;
-  deployed['contracts']['staking']['url'] = config.url + stake.address;
+  const stampABI = await require('../artifacts/src/PostageStamp.sol/PostageStamp.json');
+  deployed['contracts']['postageStamp']['abi'] = stampABI.abi;
+  deployed['contracts']['postageStamp']['bytecode'] = stampABI.bytecode.toString();
+  deployed['contracts']['postageStamp']['address'] = stamp.address;
+  deployed['contracts']['postageStamp']['block'] = deploymentReceipt.blockNumber;
+  deployed['contracts']['postageStamp']['url'] = config.url + stamp.address;
+
+  // We need to first deploy this contract and then use this address and deploy with it redistribution
+  // After that we can add here redis role
 
   // Change roles on current staking contract
   const redistributorRole = ethers.utils.keccak256(ethers.utils.toUtf8Bytes('REDISTRIBUTOR_ROLE'));
-  const tx2 = await stake.grantRole(redistributorRole, currentRedis);
+  const tx2 = await stamp.grantRole(redistributorRole, currentRedis);
   console.log('Changed REDISTRIBUTOR ROLE at : ', tx2.hash);
+
+  const oracleRole = ethers.utils.keccak256(ethers.utils.toUtf8Bytes('PRICE_ORACLE_ROLE'));
+  const tx3 = await stamp.grantRole(oracleRole, currentOracle);
+  console.log('Changed ORACLE ROLE at : ', tx3.hash);
 
   fs.writeFileSync(config.networkName + '_deployed.json', JSON.stringify(deployed, null, '\t'));
 
   if ((process.env.MAINNET_ETHERSCAN_KEY || process.env.TESTNET_ETHERSCAN_KEY) && network.name != 'localhost') {
     console.log('Verifying...');
-    await verify(stake.address, args);
+    await verify(stamp.address, args);
   }
 }
